@@ -1,10 +1,10 @@
 plugins {
-    id 'org.jetbrains.kotlin.jvm' version '1.8.0'
-    id 'org.jetbrains.kotlin.plugin.spring' version '1.8.0'
-    id 'org.jetbrains.kotlin.plugin.jpa' version '1.8.0'
+    id 'org.jetbrains.kotlin.jvm' version '1.8.22'
+    id 'org.jetbrains.kotlin.plugin.spring' version '1.8.22'
+    id 'org.jetbrains.kotlin.plugin.jpa' version '1.8.22'
 
-    id 'org.springframework.boot' version '2.6.2'
-    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
+    id 'org.springframework.boot' version '3.1.1'
+    id 'io.spring.dependency-management' version '1.1.0'
 }
 
 group = 'me.brightspark'
@@ -12,7 +12,7 @@ version = '0.1.0'
 
 kotlin {
     jvmToolchain {
-        languageVersion.set(JavaLanguageVersion.of(11))
+        languageVersion.set(JavaLanguageVersion.of(17))
     }
 }
 
@@ -31,17 +31,21 @@ repositories {
 dependencies {
     implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
     implementation 'org.jetbrains.kotlin:kotlin-reflect'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.6.4'
+    // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core-jvm
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.7.1'
 
     implementation 'org.springframework.boot:spring-boot-starter-cache'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
     runtimeOnly 'com.h2database:h2'
-    implementation 'org.xerial:sqlite-jdbc:3.39.3.0'
-    // https://github.com/gwenn/sqlite-dialect
-    implementation 'com.github.gwenn:sqlite-dialect:0.1.2'
+    // https://mvnrepository.com/artifact/org.xerial/sqlite-jdbc
+    implementation 'org.xerial:sqlite-jdbc:3.42.0.0'
+    // https://mvnrepository.com/artifact/org.hibernate.orm/hibernate-community-dialects
+    implementation 'org.hibernate.orm:hibernate-community-dialects'
 
-    implementation 'com.kotlindiscord.kord.extensions:kord-extensions:1.5.6-forums-SNAPSHOT'
-    implementation 'com.github.ben-manes.caffeine:caffeine:3.1.1'
-    implementation 'io.github.microutils:kotlin-logging-jvm:2.1.23'
+    implementation 'com.kotlindiscord.kord.extensions:kord-extensions:1.5.8-SNAPSHOT'
+    // https://mvnrepository.com/artifact/com.github.ben-manes.caffeine/caffeine
+    implementation 'com.github.ben-manes.caffeine:caffeine:3.1.6'
+    // https://mvnrepository.com/artifact/io.github.oshai/kotlin-logging-jvm
+    implementation 'io.github.oshai:kotlin-logging-jvm:4.0.0'
 }
