@@ -1,9 +1,9 @@
 package me.brightspark.mdcbot.database.model
 
-import javax.persistence.Column
-import javax.persistence.Entity
-import javax.persistence.Id
-import javax.persistence.Table
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.Id
+import jakarta.persistence.Table
 
 @Entity(name = "dev_channel")
 @Table(name = "dev_channels")
